@@ -12,7 +12,7 @@
     "clean": "",
     "code-style": "node ./just-scripts.js code-style"
   },
-  "devDependencies": {
+  "dependencies": {
     "@babel/core": "^7.8.3",
     "@babel/plugin-proposal-class-properties": "^7.7.4",
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.7.4",
@@ -25,7 +25,7 @@
     "@babel/preset-typescript": "^7.7.7",
     "@babel/register": "^7.7.7",
     "@babel/standalone": "^7.7.7",
-    "@fluentui/eslint-plugin": "*",
+    "@fluentui/eslint-plugin": "^0.50.0",
     "@mdx-js/loader": "^1.5.5",
     "@microsoft/api-extractor": "7.7.1",
     "@microsoft/load-themed-styles": "^1.10.26",
@@ -57,9 +57,6 @@
     "@types/webpack-dev-middleware": "^2.0.1",
     "@types/webpack-env": "1.15.1",
     "@types/webpack-hot-middleware": "^2.16.3",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/experimental-utils": "^2.34.0",
-    "@typescript-eslint/parser": "^2.34.0",
     "async": "^2.6.1",
     "autoprefixer": "^9.7.6",
     "babel-jest": "^24.5.0",
@@ -78,14 +75,7 @@
     "doctrine": "^2.0.0",
     "enzyme": "~3.10.0",
     "enzyme-adapter-react-16": "^1.15.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.9.0",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^4.0.0",
+    "eslint": "^7.1.0",
     "express": "^4.15.4",
     "extract-comments": "^1.0.0",
     "find-free-port": "~2.0.0",
