@@ -1128,9 +1128,20 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     };
     const { disabled } = this.props;
 
-    const handleInputKeyDown = (e: React.KeyboardEvent, searchInputProps: DropdownSearchInputProps) => {
+    const handleInputKeyDown = (
+      e: React.KeyboardEvent<HTMLInputElement>,
+      searchInputProps: DropdownSearchInputProps,
+    ) => {
       if (!disabled) {
         switch (getCode(e)) {
+          // https://github.com/downshift-js/downshift/issues/1097
+          // Downshift skips Home/End if Deopdown is opened
+          case keyboardKey.Home:
+            e.nativeEvent['preventDownshiftDefault'] = this.state.filteredItems.length === 0;
+            break;
+          case keyboardKey.End:
+            e.nativeEvent['preventDownshiftDefault'] = this.state.filteredItems.length === 0;
+            break;
           case keyboardKey.Tab:
             e.stopPropagation();
             this.handleTabSelection(e, highlightedIndex, selectItemAtIndex, toggleMenu);
@@ -1177,10 +1188,15 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
           onKeyDown: e => {
             handleInputKeyDown(e, predefinedProps);
           },
-          onChange: e => {
+          onChange: (e: React.ChangeEvent<HTMLInputElement>) => {
             // we prevent the onChange input event to bubble up to our Dropdown handler,
             // since in Dropdown it gets handled as onSearchQueryChange.
             e.stopPropagation();
+
+            // A state modification should be triggered there otherwise it will go to an another frame and will break
+            // cursor position:
+            // https://github.com/facebook/react/issues/955#issuecomment-469352730
+            this.setState({ searchQuery: e.target.value });
           },
         }),
       },
@@ -1201,7 +1217,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
       onInputBlur: (e: React.FocusEvent, searchInputProps: DropdownSearchInputProps) => {
         handleInputBlur(e, searchInputProps);
       },
-      onInputKeyDown: (e: React.KeyboardEvent, searchInputProps: DropdownSearchInputProps) => {
+      onInputKeyDown: (e: React.KeyboardEvent<HTMLInputElement>, searchInputProps: DropdownSearchInputProps) => {
         handleInputKeyDown(e, searchInputProps);
       },
     };
