@@ -80,7 +80,8 @@ const renderComponent = <P extends {}>(
     className,
     disableAnimations: context.disableAnimations || false,
     primaryDisplayName: displayName,
-    props: stateAndProps,
+    componentProps: stateAndProps,
+    inlineStylesProps: stateAndProps,
     renderer: context.renderer || noopRenderer,
     rtl,
     saveDebug,
