@@ -57,8 +57,7 @@ export const PeoplePickerProcessSelectionExample: React.FunctionComponent = () =
   };
 
   const returnMostRecentlyUsed = (currentPersonas: IPersonaProps[]): IPersonaProps[] | Promise<IPersonaProps[]> => {
-    setMostRecentlyUsed(removeDuplicates(mostRecentlyUsed, currentPersonas));
-    return filterPromise(mostRecentlyUsed);
+    return filterPromise(removeDuplicates(mostRecentlyUsed, currentPersonas));
   };
 
   const onRemoveSuggestion = (item: IPersonaProps): void => {
@@ -137,7 +136,10 @@ function listContainsPersona(persona: IPersonaProps, personas: IPersonaProps[]) 
   if (!personas || !personas.length || personas.length === 0) {
     return false;
   }
-  return personas.filter(item => item.text === persona.text).length > 0;
+  return (
+    personas.filter(item => item.text !== undefined && persona.text !== undefined && item.text.startsWith(persona.text))
+      .length > 0
+  );
 }
 
 function convertResultsToPromise(results: IPersonaProps[]): Promise<IPersonaProps[]> {
