@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ComponentProps, ShorthandValue, ComposeStandardStatics, RecursivePartial } from '../../utils/tempTypes';
-import { ComposeOptions } from '@fluentui/react-compose';
+import { BaseSlots, ComposeOptions, SlotProps } from '@fluentui/react-compose';
 import { ColorPlateSet } from '@fluentui/react-theme-provider';
 
 export type SizeValue = 'smallest' | 'smaller' | 'small' | 'medium' | 'large' | 'larger' | 'largest';
@@ -75,15 +75,13 @@ export interface ButtonProps extends ComponentProps, React.HTMLAttributes<HTMLBu
 
 export interface ButtonState extends ButtonProps {}
 
-export interface ButtonSlots {
+export interface ButtonSlots extends BaseSlots {
   icon: React.ElementType;
   content: React.ElementType;
   loader: React.ElementType;
 }
 
-export type ButtonSlotProps = {
-  [key in keyof ButtonSlots]: ButtonProps[key];
-};
+export type ButtonSlotProps = SlotProps<ButtonSlots, ButtonProps, React.ButtonHTMLAttributes<HTMLButtonElement>>;
 
 export interface ButtonOptions
   extends ComposeOptions<ButtonProps, ButtonSlots, ButtonSlotProps, ComposeStandardStatics> {}
@@ -99,6 +97,8 @@ export type ButtonTokens = ColorPlateSet & {
   iconSize: string;
   borderRadius: string;
   borderWidth: string;
+  boxShadow: string;
+  width: string;
 
   size: {
     smallest: string;
