@@ -1,12 +1,12 @@
 import * as React from 'react';
-import { ButtonProps } from './Button.types';
+import { ButtonProps, ButtonSlots, ButtonSlotProps } from './Button.types';
 import { compose, mergeProps } from '@fluentui/react-compose';
 import { useButton } from './useButton';
 
 export const ButtonBase = compose<'button', ButtonProps, ButtonProps, {}, {}>(
   (props, ref, options) => {
     const { state } = options;
-    const { slots, slotProps } = mergeProps(state, options);
+    const { slots, slotProps } = mergeProps<ButtonProps, ButtonProps, ButtonSlots, ButtonSlotProps>(state, options);
 
     return (
       <slots.root ref={ref} {...slotProps.root}>
@@ -22,7 +22,6 @@ export const ButtonBase = compose<'button', ButtonProps, ButtonProps, {}, {}>(
     handledProps: [
       'circular',
       'content',
-      'disabled',
       'fluid',
       'iconOnly',
       'iconPosition',
@@ -32,6 +31,7 @@ export const ButtonBase = compose<'button', ButtonProps, ButtonProps, {}, {}>(
       'primary',
       'secondary',
       'size',
+      'tokens',
     ],
     slots: {
       icon: 'span',
