@@ -1,5 +1,6 @@
+import * as React from 'react';
 import { ComposeStandardStatics } from '../../utils/tempTypes';
-import { ComposeOptions } from '@fluentui/react-compose';
+import { ComposeOptions, SlotProps } from '@fluentui/react-compose';
 import { ButtonProps, ButtonSlots } from '../Button/Button.types';
 
 export interface ToggleButtonProps extends ButtonProps {
@@ -22,9 +23,11 @@ export interface ToggleButtonState extends ToggleButtonProps {}
 
 export interface ToggleButtonSlots extends ButtonSlots {}
 
-export type ToggleButtonSlotProps = {
-  [key in keyof ToggleButtonSlots]: ToggleButtonProps[key];
-};
+export type ToggleButtonSlotProps = SlotProps<
+  ToggleButtonSlots,
+  ToggleButtonProps,
+  React.ButtonHTMLAttributes<HTMLButtonElement>
+>;
 
 export interface ToggleButtonOptions
   extends ComposeOptions<ToggleButtonProps, ToggleButtonSlots, ToggleButtonSlotProps, ComposeStandardStatics> {}
