@@ -1,3 +1,6 @@
-const { preset } = require('@uifabric/build');
+const { preset, just } = require('@uifabric/build');
+const { task } = just;
 
 preset();
+
+task('build', 'build:node-lib').cached();
