@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { BaseSlots, SlotProp } from '@fluentui/react-compose';
 import { IStyle, ITheme } from '../../Styling';
 import { IRefObject, IRenderFunction, IStyleFunctionOrObject } from '../../Utilities';
 import { IIconProps } from '../../Icon';
@@ -55,7 +56,7 @@ export interface ICheckboxProps extends React.ButtonHTMLAttributes<HTMLElement |
   /**
    * Label to display next to the checkbox.
    */
-  label?: string;
+  label?: SlotProp<React.HTMLAttributes<HTMLSpanElement>>;
 
   /**
    * Disabled state of the checkbox.
@@ -120,6 +121,8 @@ export interface ICheckboxProps extends React.ButtonHTMLAttributes<HTMLElement |
 
   /**
    * Custom render function for the label.
+   *
+   * @deprecated Use label prop instead.
    */
   onRenderLabel?: IRenderFunction<ICheckboxProps>;
 
@@ -203,8 +206,7 @@ export interface ICheckboxStyles {
 /**
  * {@docCategory Checkbox}
  */
-export interface ICheckboxSlots {
-  root: React.ElementType;
+export interface ICheckboxSlots extends BaseSlots {
   input: React.ElementType;
   container: React.ElementType;
   checkbox: React.ElementType;
