@@ -1,18 +1,11 @@
-let { createConfig, resolveMergeStylesSerializer } = require('@uifabric/build/jest/jest-resources');
-let path = require('path');
+const { createConfig } = require('@uifabric/build/jest/jest-resources');
 
 const config = createConfig({
-  setupFiles: [path.resolve(path.join(__dirname, 'config', 'tests.js'))],
-
   moduleNameMapper: {
-    'office-ui-fabric-react/lib/(.*)$': 'office-ui-fabric-react/lib-commonjs/$1',
-    '@uifabric/monaco-editor/lib/(.*)$': '@uifabric/monaco-editor/lib-commonjs/$1',
-    '@uifabric/react-hooks/lib/(.*)$': '@uifabric/react-hooks/lib-commonjs/$1',
-    '@uifabric/utilities/lib/(.*)$': '@uifabric/utilities/lib-commonjs/$1',
     'react-syntax-highlighter/dist/esm/(.*)$': 'react-syntax-highlighter/dist/cjs/$1',
+    // have to manually add this one because monaco-editor intentionally doesn't have a `main`
+    '@uifabric/monaco-editor/lib/(.*)$': '@uifabric/monaco-editor/lib-commonjs/$1',
   },
-
-  snapshotSerializers: [resolveMergeStylesSerializer()],
 });
 
 module.exports = config;
