@@ -1,4 +1,5 @@
-const { preset } = require('@uifabric/build');
+const { preset, just } = require('@uifabric/build');
+const { task } = just;
 
 preset();
 
