@@ -1,7 +1,39 @@
+// @ts-check
+
+const fs = require('fs-extra');
 const path = require('path');
 const merge = require('../tasks/merge');
 const resolve = require('resolve');
 const { resolveCwd } = require('just-scripts');
+const { findRepoDeps } = require('../monorepo/index');
+const findConfig = require('../find-config');
+
+const packageJsonPath = findConfig('package.json');
+const packageRoot = path.dirname(packageJsonPath);
+
+const jestAliases = () => {
+  // Get deps of the current package within the repo
+  const packageDeps = findRepoDeps();
+
+  const aliases = {};
+
+  for (const { packageJson } of packageDeps) {
+    const { name, main } = packageJson;
+    if (main && main.includes('lib-commonjs')) {
+      // Map package root and lib imports to the commonjs version
+      const mainImportPath = `${name}/${main.replace('.js', '')}`;
+      aliases[`^${name}$`] = mainImportPath;
+      aliases[`^${name}/lib/(.*)$`] = mainImportPath.replace('index', '$1');
+    }
+  }
+
+  // Special aliases to look at src for the current package
+  const packageJson = fs.readJSONSync(packageJsonPath);
+  aliases[`^${packageJson.name}$`] = '<rootDir>/src/';
+  aliases[`^${packageJson.name}/lib/(.*)$`] = '<rootDir>/src/$1';
+
+  return aliases;
+};
 
 module.exports = {
   resolveMergeStylesSerializer: () => resolveCwd('@uifabric/jest-serializer-merge-styles'),
@@ -9,6 +41,7 @@ module.exports = {
     rootDir: 'lib',
     testRegex: '(/__tests__/.*|\\.(test|spec))\\.js$',
   }),
+  jestAliases,
   createConfig: customConfig =>
     merge(
       {
@@ -16,6 +49,7 @@ module.exports = {
           'ts-jest': resolve.sync('ts-jest'),
           '\\.(scss)$': path.resolve(__dirname, 'jest-style-mock.js'),
           KeyCodes: path.resolve(__dirname, 'jest-mock.js'),
+          ...jestAliases(),
         },
 
         transform: {
@@ -33,14 +67,14 @@ module.exports = {
 
         moduleDirectories: [
           'node_modules',
-          path.resolve(process.cwd(), 'node_modules'),
+          path.resolve(packageRoot, 'node_modules'),
           path.resolve(__dirname, '../node_modules'),
         ],
 
         globals: {
           'ts-jest': {
-            tsConfig: path.resolve(process.cwd(), 'tsconfig.json'),
-            packageJson: path.resolve(process.cwd(), 'package.json'),
+            tsConfig: path.resolve(packageRoot, 'tsconfig.json'),
+            packageJson: path.resolve(packageRoot, 'package.json'),
             diagnostics: false,
           },
         },
