@@ -1,7 +1,11 @@
+// @ts-check
+
 const { readConfig } = require('../read-config');
-const path = require('path');
 const getAllPackageInfo = require('./getAllPackageInfo');
 
+/**
+ * @param {import('./index').PackageJson} packageJson
+ */
 function getDeps(packageJson) {
   if (!packageJson) {
     return [];
@@ -10,15 +14,24 @@ function getDeps(packageJson) {
   return Object.keys({ ...(packageJson.dependencies || {}), ...(packageJson.devDependencies || {}) });
 }
 
+/** @type {import('./index').PackageInfo[]} */
+let repoDeps;
+let cwdForRepoDeps;
+
 /**
  * Find all the dependencies (and their dependencies) within the repo for a specific package (in the CWD when this was called)
- * @returns {Map<string, Set<string>>}
+ * @returns {import('./index').PackageInfo[]}
  */
 function findRepoDeps() {
+  const cwd = process.cwd();
+  if (repoDeps && cwdForRepoDeps === cwd) {
+    return repoDeps;
+  }
+
   const packageInfo = getAllPackageInfo();
-  let cwd = process.cwd();
-  const packageJson = readConfig(path.join(cwd, 'package.json'));
+  const packageJson = readConfig('package.json');
   const packageDeps = getDeps(packageJson);
+  /** @type {Set<string>} */
   const result = new Set();
 
   while (packageDeps.length > 0) {
@@ -35,11 +48,14 @@ function findRepoDeps() {
     }
   }
 
-  return [...result].map(dep => packageInfo[dep]);
+  repoDeps = [...result].map(dep => packageInfo[dep]);
+  cwdForRepoDeps = cwd;
+  return repoDeps;
 }
 
 module.exports = findRepoDeps;
 
+// @ts-ignore
 if (require.main === module) {
   console.log(findRepoDeps());
 }
